refactor(e-hub): arranging the folder and file
