+++ src/components/Button/button.tsx
@@ -1,0 +1,7 @@
+const Button = () => {
+    return (
+        <button>Button</button>
+    );
+};
+
+export default Button;
