+++ src/components/input/input.tsx
@@ -1,0 +1,7 @@
+const Input = () => {
+    return (
+        <p>Input</p>
+    );
+};
+
+export default Input;
